+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Changes
 - Revoked previous backend acceptance to display privacy policy
 - Added (ᵔᴥᵔ) as a trigger for the :dog: emoji
+- Made the introduction walkthrough send as a single message upon joining SkyBlock instead of separate messages on Hypixel join
 
 ### Fixes
 - Fixed sea creature glow only displaying one entity and invisible entities sometimes